+++ nest_lesson_3/test/users.e2e-spec.ts
@@ -1,28 +1,7 @@
 import { HttpStatus, INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { testInitSettings } from 'src/common/test-init-settings';
-import { faker } from '@faker-js/faker';
-
-const createUserAndLogin = async (httpServer) => {
-  const newUser = {
-    login: faker.string.alpha({ length: { min: 3, max: 10 } }),
-    email: faker.internet.email(),
-    password: faker.internet.password({ length: 12 }),
-  };
-
-  const resUsers = await request(httpServer)
-    .post('/users')
-    .set({ Authorization: 'Basic YWRtaW46cXdlcnR5' })
-    .send(newUser);
-
-  const response = await request(httpServer)
-    .post('/auth/login')
-    .send({ loginOrEmail: newUser.login, password: newUser.password });
-
-  console.log(resUsers.status);
-
-  return response.body;
-};
+import { createUserAndLogin } from './helpers/auth.helpers';
 
 describe('Users tests', () => {
   let app: INestApplication;
@@ -40,37 +19,34 @@ describe('Users tests', () => {
   });
 
   describe('Auth user tests', () => {
-    let tokens;
+    let accessToken;
+    let user;
 
     beforeAll(async () => {
-      tokens = await createUserAndLogin(httpServer);
+      const res = await createUserAndLogin(httpServer);
 
-      console.log('tokens', tokens);
+      accessToken = res.accessToken;
+      user = res.user;
     });
 
     test('user should created', () => {
       return request(httpServer)
         .post('/users')
-        .send({
-          login: faker.person.firstName(),
-          email: faker.internet.email(),
-          password: faker.internet.password({ length: 8 }),
-        })
+        .set({ Authorization: 'Basic YWRtaW46cXdlcnR5' })
+        .send(user)
         .expect(HttpStatus.CREATED);
     });
 
     test('should return correct response body', async () => {
-      const login = faker.person.firstName();
-      const email = faker.internet.email();
-      const password = faker.internet.password({ length: 8 });
+      const response = await request(httpServer)
+        .post('/users')
+        .set({ Authorization: 'Basic YWRtaW46cXdlcnR5' })
+        .send(user);
 
-      const response = await request(httpServer).post('/users').send({
-        login,
-        email,
-        password,
+      expect(response.body).toMatchObject({
+        login: user.login,
+        email: user.email,
       });
-
-      expect(response.body).toMatchObject({ login, email });
       expect(response.body).toHaveProperty('id');
       expect(response.body).toHaveProperty('createdAt');
     });
@@ -78,9 +54,10 @@ describe('Users tests', () => {
     test('should failed if login empty', () => {
       return request(httpServer)
         .post('/users')
+        .set({ Authorization: 'Basic YWRtaW46cXdlcnR5' })
         .send({
-          email: faker.internet.email(),
-          password: faker.internet.password({ length: 8 }),
+          email: user.email,
+          password: user.password,
         })
         .expect(HttpStatus.BAD_REQUEST);
     });
@@ -88,9 +65,10 @@ describe('Users tests', () => {
     test('should failed if email empty', () => {
       return request(httpServer)
         .post('/users')
+        .set({ Authorization: 'Basic YWRtaW46cXdlcnR5' })
         .send({
-          login: faker.person.firstName(),
-          password: faker.internet.password({ length: 8 }),
+          login: user.login,
+          password: user.password,
         })
         .expect(HttpStatus.BAD_REQUEST);
     });
@@ -98,9 +76,10 @@ describe('Users tests', () => {
     test('should failed if password empty', () => {
       return request(httpServer)
         .post('/users')
+        .set({ Authorization: 'Basic YWRtaW46cXdlcnR5' })
         .send({
-          login: faker.person.firstName(),
-          email: faker.internet.email(),
+          login: user.login,
+          email: user.email,
         })
         .expect(HttpStatus.BAD_REQUEST);
     });
